+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Fixed iterator closing in `Set.prototype.{ isDisjointFrom, isSupersetOf }` polyfill
 - Fixed (updated following the final spec) one more case `Set.prototype.difference` polyfill with updating `this`
 - Fixed `DataView.prototype.setFloat16` polyfill in (0, 1) range
+- Fixed order of arguments validation in `String.prototype.{ startsWith, endsWith }` polyfills
 - Fixed named backreferences in `RegExp` NCG polyfill
 - Fixed some cases of `RegExp` NCG polyfill in combination with other types of groups
 - Fixed some cases of `RegExp` NCG polyfill in combination with `dotAll`